Update only rate changes

+++ domain/src/main/java/com/wojdor/domain/Rate.kt
@@ -1,5 +1,6 @@
 package com.wojdor.domain
 
 import com.wojdor.domain.enums.Currency
+import java.math.BigDecimal
 
-data class Rate(val currency: Currency, val rate: Double)
+data class Rate(val currency: Currency, val rate: BigDecimal)
